Add delete

+++ src/components/posts/PostListItem.vue
@@ -4,8 +4,8 @@
 		<div class="post-contents">{{ postItem.contents }}</div>
 		<div class="post-time">
 			{{ postItem.createdAt }}
-			<i class="icon ion-md-create"></i>
-			<i class="icon ion-md-trash" @click=""></i>
+			<i class="icon ion-md-create" @click="routeEditPage"></i>
+			<i class="icon ion-md-trash" @click="deleteItem"></i>
 		</div>
 	</li>
 </template>
@@ -21,7 +21,14 @@ export default {
 	},
 	methods: {
 		async deleteItem() {
-			await deletePost(this.postItem._id);
+			if (confirm('want ')) {
+				await deletePost(this.postItem._id);
+				this.$emit('refresh');
+			}
+			console.log('deleted');
+		},
+		routeEditPage() {
+			this.$router.push('/');
 		},
 	},
 };
